sm_specific_phase is added to merge generic and specific phases. fom state is stored in fom::fo_sm_state::sm_rc while returning from fom phase routines.  fom_type->ft_phases[C2_FOPH_NR].sd_allowed |= (1 << (C2_FOPH_NR + 1)); is added in fom register if FOM specific phase are present. This means after generic phase transition, always first specfic transition needs to be called.
struct c2_sm_state_descr io_states contains allowed transitions and state functions
for both read and write states.
struct c2_io_fom_cob_rw_state_transition io_fom_read_st and io_fom_write_st are
removed and next phase is added in corresponding state function dependent on fop type.
When fom phase is C2_FOPH_SUCCESS state machine returns and completes remianing generic
phases.
reply_fop_populate() is added in io_foms.c to fill reply fop after
successful read/write.
typecast with uint64_t for 1 is added in ((uint64_t)1 << C2_FOPH_IO_BUFFER_RELEASE)
as value of C2_FOPH_IO_BUFFER_RELEASE is greater than 32 and is also added in state_set()
as ((sd->sd_allowed & ((uint64_t)1 << state)).

io_states are added for fom types ut_io_fom_cob_rw_type_mopt,bulkio_server_read_fom_type and
bulkio_server_write_fom_type in bulkio_ut.c and are registered using c2_fom_type_register().
Instead of c2_io_fom_cob_rw_state() corresponding state functions are used in
bulkio_server_write_fom_state() and bulkio_server_read_fom_state().

+++ sm/sm.c
@@ -238,7 +238,7 @@ static void state_set(struct c2_sm *mach, int state)
 	 */
 	do {
 		sd = sm_state(mach);
-		C2_PRE(sd->sd_allowed & (1 << state));
+		C2_PRE(sd->sd_allowed & ((uint64_t)1 << state));
 
 		if (sd->sd_ex != NULL)
 			sd->sd_ex(mach);
